+++ packages/analytics-js-plugins/src/beaconQueue/index.ts
@@ -11,7 +11,7 @@ import type {
 import type { RudderEvent } from '@rudderstack/analytics-js-common/types/Event';
 import type { ExtensionPlugin } from '@rudderstack/analytics-js-common/types/PluginEngine';
 import type { PluginName } from '@rudderstack/analytics-js-common/types/PluginsManager';
-import type { DoneCallback, IQueue } from '../types/plugins';
+import type { DoneCallback, IQueue, QueueProcessCallbackInfo } from '../types/plugins';
 import {
   getNormalizedBeaconQueueOptions,
   getDeliveryUrl,
@@ -65,29 +65,47 @@ const BeaconQueue = (): ExtensionPlugin => ({
         state.loadOptions.value.beaconQueueOptions ?? {},
       );
 
-      const queueProcessCallback = (itemData: BeaconQueueBatchItemData, done: DoneCallback) => {
+      const queueProcessCallback = (
+        itemData: BeaconQueueBatchItemData,
+        done: DoneCallback,
+        info: QueueProcessCallbackInfo,
+      ) => {
+        const { isPageAccessible } = info;
+
         logger?.debug(BEACON_PLUGIN_EVENTS_QUEUE_DEBUG(BEACON_QUEUE_PLUGIN));
         const currentTime = getCurrentTimeFormatted();
         const finalEvents = itemData.map((queueItemData: BeaconQueueItemData) =>
           getFinalEventForDeliveryMutator(queueItemData.event, currentTime),
         );
         const data = getBatchDeliveryPayload(finalEvents, currentTime, logger);
+        if (!data) {
+          // Mark the item as done so that it can be removed from the queue
+          done(null);
+          return;
+        }
 
-        if (data) {
-          try {
-            const isEnqueuedInBeacon = navigator.sendBeacon(url, data);
-            if (!isEnqueuedInBeacon) {
-              logger?.error(BEACON_QUEUE_SEND_ERROR(BEACON_QUEUE_PLUGIN));
+        try {
+          if (!navigator.sendBeacon(url, data)) {
+            if (isPageAccessible) {
+              logger?.error(
+                `${BEACON_QUEUE_SEND_ERROR(BEACON_QUEUE_PLUGIN)} The event(s) will be dropped.`,
+              );
+
+              // Remove the item from queue
+              done(null);
+            } else {
+              // Note: We're not removing the item from the queue as we want to retry the request
+              logger?.warn(
+                `${BEACON_QUEUE_SEND_ERROR(BEACON_QUEUE_PLUGIN)} The event(s) will be retried as the current page is being unloaded.`,
+              );
             }
-
-            done(null, isEnqueuedInBeacon);
-          } catch (err) {
-            errorHandler?.onError(err, BEACON_QUEUE_PLUGIN, BEACON_QUEUE_DELIVERY_ERROR(url));
-            // Remove the item from queue
+          } else {
+            // Remove the item from queue as the request was successful
             done(null);
           }
-        } else {
-          // Mark the item as done so that it can be removed from the queue
+        } catch (err) {
+          errorHandler?.onError(err, BEACON_QUEUE_PLUGIN, BEACON_QUEUE_DELIVERY_ERROR(url));
+          // Remove the item from queue
           done(null);
         }
       };