+++ packages/analytics-js/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.19.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.18.2...@rudderstack/analytics-js@3.19.0) (2025-06-03)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-plugins` updated to version `3.9.0`
+
+### Features
+
+* retry delivery failures in beacon plugin when page is unloaded ([#2269](https://github.com/rudderlabs/rudder-sdk-js/issues/2269)) ([cec81f3](https://github.com/rudderlabs/rudder-sdk-js/commit/cec81f3d2aca443f6d2c209941dd28fffd65888c))
+
 ## [3.18.2](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.18.1...@rudderstack/analytics-js@3.18.2) (2025-05-26)
 
 ### Dependency Updates